Remove Response Base and status code

+++ BackEnd/Controllers/Equipments/EquipmentController.cs
@@ -5,30 +5,23 @@ using System.Threading.Tasks;
 using AutoMapper;
 using AutoMapper.QueryableExtensions;
 using BackEnd.DataBase;
-using BackEnd.Exceptions;
 using BackEnd.Models.Roles;
 using Extensions;
-using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Identity;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.EntityFrameworkCore;
 using Microsoft.Extensions.Logging;
-using Models;
 using Models.Equipments;
 using Models.People;
 using Models.People.Roles;
 using Models.PublicAPI.Requests;
 using Models.PublicAPI.Requests.Equipment.Equipment;
-using Models.PublicAPI.Responses;
 using Models.PublicAPI.Responses.Equipment;
-using Models.PublicAPI.Responses.General;
-using Microsoft.Data.Edm.Csdl;
-using BackEnd.Extensions;
 using System.Threading;
 
 namespace BackEnd.Controllers.Equipments
 {
-    [Produces("application/json")]
+
     [Route("api/Equipment")]
     public class EquipmentController : AuthorizeController
     {
@@ -51,15 +44,19 @@ namespace BackEnd.Controllers.Equipments
 
 
         [HttpGet("{id}")]
-        public async Task<OneObjectResponse<EquipmentView>> GetAsync(Guid id)
-            => await dbContext
-                .Equipments
-                .ProjectTo<EquipmentView>()
-                .SingleOrDefaultAsync(eq => eq.Id == id)
-                ?? throw NotFoundMyApi();
+        public async Task<ActionResult<EquipmentView>> GetAsync(Guid id)
+        {
+            var equipmentView = await dbContext
+                                      .Equipments
+                                      .ProjectTo<EquipmentView>()
+                                      .SingleOrDefaultAsync(eq => eq.Id == id);
+            if (equipmentView == null)
+                return NotFound();
+            return equipmentView;
+        }
 
         [HttpGet]
-        public async Task<ListResponse<CompactEquipmentView>> GetAsync(
+        public async Task<ActionResult<List<CompactEquipmentView>>> GetAsync(
             Guid? eventId,
             Guid? equipmentTypeId,
             string match,
@@ -71,7 +68,7 @@ namespace BackEnd.Controllers.Equipments
                          eq => eq.ParentId == null)
                 .WhereIf(eventId.HasValue, eq => eq.PlaceEquipments.Any(pe => pe.Place.Shift.EventId == eventId))
                 .WhereIf(equipmentTypeId.HasValue, eq => eq.EquipmentTypeId == equipmentTypeId)
-                
+
                 .IfNotNull(match, equipments => equipments
                     .Variable(out var tokens, () => match.Split(' '))
                     .Variable(out var ints, () => tokens.GetInts())
@@ -83,16 +80,33 @@ namespace BackEnd.Controllers.Equipments
                 .ProjectTo<CompactEquipmentView>()
                 .ToListAsync();
 
+        /// <summary>
+        /// Create new Equipment
+        /// </summary>
+        /// <param name="request"></param>
+        /// <returns>A newly created Equipment</returns>
+        /// <response code="200">Returns newly created Equipment</response>
+        /// <response code="400">If incorrect children id passed</response>
+        /// <response code="404">If equipment type id don't exist</response>
+        /// <response code="409">If serial number exists</response>
         [RequireRole(RoleNames.CanEditEquipment)]
         [HttpPost]
-        public async Task<OneObjectResponse<EquipmentView>> PostAsync([FromBody]EquipmentCreateRequest request)
+        [ProducesResponseType(200)]
+        [ProducesResponseType(400)]
+        [ProducesResponseType(404)]
+        [ProducesResponseType(409)]
+        public async Task<ActionResult<EquipmentView>> PostAsync([FromBody]EquipmentCreateRequest request)
         {
             try
             {
                 await semaphore.WaitAsync();
 
-                var type = await CheckAndGetEquipmentTypeAsync(request.EquipmentTypeId);
-                await CheckNotExist(request.SerialNumber);
+                var type = await dbContext.EquipmentTypes.FindAsync(request.EquipmentTypeId);
+                if (type == null)
+                    return NotFound();
+
+                if (!await CheckNotExist(request.SerialNumber))
+                    return Conflict("Serial number exists");
 
                 var newEquipment = mapper.Map<Equipment>(request);
                 newEquipment.Number = type.LastNumber++;
@@ -104,7 +118,7 @@ namespace BackEnd.Controllers.Equipments
                         .ToListAsync();
 
                 if (newEquipment.Children?.Count != request.Children?.Count)
-                    throw ResponseStatusCode.IncorrectEquipmentIds.ToApiException();
+                    return BadRequest("Incorrect equipment ids");
 
                 await dbContext.Equipments.AddAsync(newEquipment);
                 await dbContext.SaveChangesAsync();
@@ -119,13 +133,17 @@ namespace BackEnd.Controllers.Equipments
 
         [RequireRole(RoleNames.CanEditEquipment)]
         [HttpPut]
-        public async Task<OneObjectResponse<EquipmentView>> PutAsync(int id, [FromBody]EquipmentEditRequest request)
+        public async Task<ActionResult<EquipmentView>> PutAsync(int id, [FromBody]EquipmentEditRequest request)
         {
             var toEdit = await CheckAndGetEquipmentAsync(request.Id);
             if (request.EquipmentTypeId.HasValue)
-                await CheckAndGetEquipmentTypeAsync(request.EquipmentTypeId.Value);
+            {
+                if (!await dbContext.EquipmentTypes.AnyAsync(et => et.Id == request.EquipmentTypeId.Value))
+                    return NotFound();
+            }
             if (string.IsNullOrEmpty(request.SerialNumber))
-                await CheckNotExist(request.SerialNumber);
+                if (!await CheckNotExist(request.SerialNumber))
+                    return Conflict("Serial number exists");//TODO meta
 
             mapper.Map(request, toEdit);
             await dbContext.SaveChangesAsync();
@@ -135,29 +153,27 @@ namespace BackEnd.Controllers.Equipments
 
         [RequireRole(RoleNames.CanEditEquipment)]
         [HttpDelete]
-        public async Task<OneObjectResponse<Guid>> DeleteAsync([FromBody]IdRequest request)
+        public async Task<ActionResult<Guid>> DeleteAsync([FromBody]IdRequest request)
         {
             var toDelete = await CheckAndGetEquipmentAsync(request.Id);
+            if (toDelete == null)
+                return NotFound();
             dbContext.Equipments.Remove(toDelete);
             await dbContext.SaveChangesAsync();
             return request.Id;
         }
 
-        private async Task<Equipment> CheckAndGetEquipmentAsync(Guid id)
-            => await dbContext.Equipments.Include(eq => eq.EquipmentType).FirstOrDefaultAsync(eq => eq.Id == id)
-              ?? throw NotFoundMyApi();
-
-        private async Task CheckNotExist(string serialNumber)
+        private Task<Equipment> CheckAndGetEquipmentAsync(Guid id)
         {
-            if (await dbContext
-                .Equipments
-                .AnyAsync(eq => eq.SerialNumber == serialNumber))
-                throw ApiLogicException.Create(ResponseStatusCode.FieldExist);
+            return dbContext.Equipments.Include(eq => eq.EquipmentType).FirstOrDefaultAsync(eq => eq.Id == id);
         }
 
-        private async Task<EquipmentType> CheckAndGetEquipmentTypeAsync(Guid typeId)
-            => await dbContext.EquipmentTypes.FindAsync(typeId)
-                ?? throw ApiLogicException.Create(ResponseStatusCode.EquipmentTypeNotFound);
+        private Task<bool> CheckNotExist(string serialNumber)
+        {
+            return dbContext
+                .Equipments
+                .AnyAsync(eq => eq.SerialNumber == serialNumber);
+        }
 
     }
 }
